created kenlm_text_utils.py
